requirements/development.txt
-r base.txt

django-debug-toolbar==4.3.0
flake8==7.0.0
ipython==8.24.0
isort==5.13.2